feat: Update Conversation type to include unreadMessagesCount
This commit updates the Conversation type to include the unreadMessagesCount property. This property will be used to keep track of the number of unread messages in a conversation. This change is necessary to support the new feature of displaying the number of new messages in the conversation item and the chat view.